rename has/has to core/has
