+++ docs/core/has.md
@@ -2,11 +2,11 @@
 
 ## Detecting Features
 
-The default export of `@dojo/framework/has/has` is a function which accepts a single parameter: the name of the feature to test for.
+The default export of `@dojo/framework/core/has` is a function which accepts a single parameter: the name of the feature to test for.
 If the feature is available, a truthy value is returned, otherwise a falsy value is returned:
 
 ```ts
-import has from '@dojo/framework/has/has';
+import has from '@dojo/framework/core/has';
 
 if (has('dom-addeventlistener')) {
 	element.addEventListener('click', function() {
@@ -17,7 +17,7 @@ if (has('dom-addeventlistener')) {
 
 ## Adding Feature Detections
 
-It's important to be able to add new feature tests that aren't provided out-of-the-box by `@dojo/framework/has/has`.
+It's important to be able to add new feature tests that aren't provided out-of-the-box by `@dojo/framework/core/has`.
 This can be done easily by using the `add` function exported by the `has` module. It accepts two parameters:
 the name of the feature, and either an immediate value indicating its availability or a function that resolves to a
 value.
@@ -26,7 +26,7 @@ When a function is passed, the feature will be lazily evaluated - i.e. the funct
 actually requested. The return value is then cached for future calls for the same feature.
 
 ```ts
-import { add as hasAdd } from '@dojo/framework/has/has';
+import { add as hasAdd } from '@dojo/framework/core/has';
 hasAdd('dom-queryselector', 'querySelector' in document && 'querySelectorAll' in document);
 
 // Lazily executed; useful if a polyfill is loaded after page load
@@ -37,6 +37,6 @@ hasAdd('typedarray', function() {
 
 ## Accessing the Feature Cache
 
-`@dojo/framework/has/has` maintains object hashes containing keys that correspond to all features that have been both
+`@dojo/framework/core/has` maintains object hashes containing keys that correspond to all features that have been both
 registered _and_ requested. The value associated with each feature name key corresponds to that feature's availability
 in the current environment. The object hash containing evaluated features is accessible via the `cache` export.
