+++ src/shim/Map.ts
@@ -1,7 +1,7 @@
 import { isArrayLike, Iterable, IterableIterator, ShimIterator } from './iterator';
 import global from './global';
 import { is as objectIs } from './object';
-import has from '../has/has';
+import has from '../core/has';
 import './Symbol';
 
 export interface Map<K, V> {
