+++ tests/core/support/sendEvent.ts
@@ -1,4 +1,4 @@
-import has, { add } from '../../../src/has/has';
+import has, { add } from '../../../src/core/has';
 import { deepAssign } from '../../../src/core/util';
 import global from '../../../src/shim/global';
 import { assign } from '../../../src/shim/object';
